chore: move final task into tasks
